@@ -59,7 +59,7 @@ class ServiceLoader implements BlockLoaderInterface
         $block->setEnabled(true);
         $block->setCreatedAt(new \DateTime());
         $block->setUpdatedAt(new \DateTime());
-        $block->setSettings(isset($configuration['settings']) ? $configuration['settings'] : array());
+        $block->setSettings(isset($configuration['settings']) ? $configuration['settings'] : []);
 
         return $block;
     }
