@@ -70,8 +70,8 @@ class ContainerTemplateType extends AbstractType
      */
     public function configureOptions(OptionsResolver $resolver)
     {
-        $resolver->setDefaults(array(
+        $resolver->setDefaults([
             'choices' => $this->templateChoices,
-        ));
+        ]);
     }
 }
