@@ -11,10 +11,10 @@
 
 namespace Sonata\BlockBundle\Twig\Extension;
 
+use PHPUnit\Framework\TestCase;
 use Sonata\BlockBundle\Templating\Helper\BlockHelper;
-use Sonata\BlockBundle\Tests\PHPUnit_Framework_TestCase;
 
-class BlockExtensionTest extends PHPUnit_Framework_TestCase
+class BlockExtensionTest extends TestCase
 {
     /**
      * @var \PHPUnit_Framework_MockObject_MockObject|BlockHelper
@@ -47,20 +47,20 @@ class BlockExtensionTest extends PHPUnit_Framework_TestCase
 
     public function provideFunction()
     {
-        return array(
-            array('sonata_block_render', array(
-                'foobar', array('bar' => 'foo'),    // arguments
-            ), 'render'),
-            array('sonata_block_include_javascripts', array(
+        return [
+            ['sonata_block_render', [
+                'foobar', ['bar' => 'foo'],    // arguments
+            ], 'render'],
+            ['sonata_block_include_javascripts', [
                 'screen',                         // arguments
-            ), 'includeJavascripts'),
-            array('sonata_block_include_stylesheets', array(
+            ], 'includeJavascripts'],
+            ['sonata_block_include_stylesheets', [
                 'foo',                            // arguments
-            ), 'includeStylesheets'),
-            array('sonata_block_render_event', array(
-                'event.name', array(),            // arguments
-            ), 'renderEvent'),
-        );
+            ], 'includeStylesheets'],
+            ['sonata_block_render_event', [
+                'event.name', [],            // arguments
+            ], 'renderEvent'],
+        ];
     }
 
     /**
